Get the design working with Vue.js
I converted some files from dos to unix, so you need to ignore
whitespace when viewing this diff.

This is a functional webpage with vue mixed in.  Took a lot of work to
get the jquery code to work well with vue.  The code was storing lots of
dom in variables to use later, so I converted it to look up the dom each
time instead of storing it.

It's now single page.  I'll delete the other orphaned pages in a later
commit.

Removed scrolly library and added a vue scrolling library.  When
clicking a link to another "page" on the site, since it's single-page
now, I make it scroll to the header. Otherwise when navagating, it would
leave the scroll position in the same space, which might be way down the
page.

The resume page still needs a lot of work, and the content is not final,
but it's getting closer